requirements.txt
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0
opencv-python>=4.5.0
tensorflow>=2.8.0
numpy>=1.21.0
Pillow>=8.3.0
streamlit>=1.28.0
flask>=2.3.0
flask-socketio>=5.3.0